requirements_bridge.txt
aiohttp>=3.8.0
paho-mqtt>=1.6.0
requests>=2.28.0